intro-flask/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
filelock==3.12.0
Flask==2.2.1
fonttools==4.38.0
future==0.18.2
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mysql-connector-python==8.0.30
numpy==1.23.4
packaging==23.0
Pillow==9.4.0
platformdirs==3.5.1
protobuf==3.20.1
pycodestyle==2.10.0
pycryptodome==3.18.0
pygame==2.1.2
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
PyYAML==6.0
serial==0.0.97
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
virtualenv==20.23.0
Werkzeug==2.2.1
